pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hp-sku-photo-scraper"
version = "0.1.0"
description = "Download product images for HP SKUs from HP Shop search results"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "requests>=2.31",
  "selenium>=4.15",
  "pillow>=10.0",
  "webdriver-manager>=4.0.0",
  "tqdm>=4.66",
]

[project.scripts]
hp-sku-scraper = "hp_sku_photo_scraper.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
